requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
asgiref==3.7.2
    # via
    #   -r requirements.in
    #   django
build==0.10.0
    # via
    #   -r requirements.in
    #   pip-tools
cachetools==5.3.1
    # via
    #   -r requirements.in
    #   google-auth
certifi==2023.7.22
    # via
    #   -r requirements.in
    #   requests
cffi==1.15.1
    # via
    #   -r requirements.in
    #   cryptography
charset-normalizer==2.1.1
    # via
    #   -r requirements.in
    #   requests
click==8.1.6
    # via
    #   -r requirements.in
    #   pip-tools
cryptography==41.0.3
    # via
    #   -r requirements.in
    #   social-auth-core
defusedxml==0.7.1
    # via
    #   -r requirements.in
    #   python3-openid
    #   social-auth-core
django==4.1.5
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   djoser
    #   drf-yasg
    #   social-auth-app-django
django-cors-headers==4.2.0
    # via -r requirements.in
django-templated-mail==1.1.1
    # via
    #   -r requirements.in
    #   djoser
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework-simplejwt==5.2.2
    # via
    #   -r requirements.in
    #   djoser
djoser==2.2.0
    # via -r requirements.in
drf-yasg==1.21.7
    # via -r requirements.in
google-auth==2.22.0
    # via
    #   -r requirements.in
    #   google-auth-oauthlib
    #   gspread
google-auth-oauthlib==1.0.0
    # via
    #   -r requirements.in
    #   gspread
gspread==5.7.2
    # via -r requirements.in
httplib2==0.22.0
    # via
    #   -r requirements.in
    #   oauth2client
idna==3.4
    # via
    #   -r requirements.in
    #   requests
inflection==0.5.1
    # via
    #   -r requirements.in
    #   drf-yasg
oauth2client==4.1.3
    # via -r requirements.in
oauthlib==3.2.2
    # via
    #   -r requirements.in
    #   requests-oauthlib
    #   social-auth-core
packaging==23.1
    # via
    #   -r requirements.in
    #   build
    #   drf-yasg
pip-tools==7.3.0
    # via -r requirements.in
pyasn1==0.5.0
    # via
    #   -r requirements.in
    #   oauth2client
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via
    #   -r requirements.in
    #   google-auth
    #   oauth2client
pycparser==2.21
    # via
    #   -r requirements.in
    #   cffi
pyjwt==2.8.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   social-auth-core
pyparsing==3.1.1
    # via
    #   -r requirements.in
    #   httplib2
pyproject-hooks==1.0.0
    # via
    #   -r requirements.in
    #   build
python-dotenv==1.0.0
    # via -r requirements.in
python3-openid==3.2.0
    # via
    #   -r requirements.in
    #   social-auth-core
pytz==2023.3
    # via
    #   -r requirements.in
    #   djangorestframework
    #   drf-yasg
pyyaml==6.0.1
    # via
    #   -r requirements.in
    #   drf-yasg
requests==2.28.1
    # via
    #   -r requirements.in
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==1.3.1
    # via
    #   -r requirements.in
    #   google-auth-oauthlib
    #   social-auth-core
rsa==4.9
    # via
    #   -r requirements.in
    #   google-auth
    #   oauth2client
six==1.16.0
    # via
    #   -r requirements.in
    #   google-auth
    #   oauth2client
social-auth-app-django==5.2.0
    # via
    #   -r requirements.in
    #   djoser
social-auth-core==4.4.2
    # via
    #   -r requirements.in
    #   social-auth-app-django
sqlparse==0.4.4
    # via
    #   -r requirements.in
    #   django
tomli==2.0.1
    # via
    #   -r requirements.in
    #   build
    #   pip-tools
    #   pyproject-hooks
typing-extensions==4.7.1
    # via
    #   -r requirements.in
    #   asgiref
uritemplate==4.1.1
    # via
    #   -r requirements.in
    #   drf-yasg
urllib3==1.26.16
    # via
    #   -r requirements.in
    #   google-auth
    #   requests
wheel==0.41.1
    # via pip-tools
xlsxwriter==3.1.2
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
